test-LeNet-5/requirements.txt
Package                Version
---------------------- ---------
-eras-Applications     1.0.8
absl-py                0.11.0
astor                  0.8.1
astunparse             1.6.3
blis                   0.4.1
cachetools             4.1.1
catalogue              1.0.0
certifi                2020.6.20
chardet                3.0.4
cycler                 0.10.0
cymem                  2.0.4
decorator              4.4.2
gast                   0.3.3
google-auth            1.23.0
google-auth-oauthlib   0.4.2
google-pasta           0.2.0
grpcio                 1.33.2
h5py                   2.8.0
idna                   2.10
imageio                2.9.0
imgaug                 0.4.0
importlib-metadata     2.0.0
inference              0.1
Keras                  2.2.2
Keras-Applications     1.0.4
keras-models           0.0.7
Keras-Preprocessing    1.0.2
kiwisolver             1.3.1
Markdown               3.3.3
matplotlib             3.3.2
mock                   4.0.2
murmurhash             1.0.4
networkx               2.5
numpy                  1.15.1
oauthlib               3.1.0
opencv-python          4.4.0.46
opt-einsum             3.3.0
pathlib                1.0.1
Pillow                 8.0.1
pip                    20.2.4
plac                   1.1.3
preshed                3.0.4
protobuf               3.13.0
pyasn1                 0.4.8
pyasn1-modules         0.2.8
pyparsing              2.4.7
python-dateutil        2.8.1
PyWavelets             1.1.1
PyYAML                 5.3.1
requests               2.24.0
requests-oauthlib      1.3.0
rsa                    4.6
scikit-image           0.17.2
scipy                  1.5.3
setuptools             50.3.2
Shapely                1.7.1
six                    1.11.0
spacy                  2.3.2
srsly                  1.0.3
tensorboard            1.13.1
tensorboard-plugin-wit 1.7.0
tensorflow             1.13.1
tensorflow-estimator   1.13.0
tensorflow-hub         0.1.1
tensorflowjs           0.8.5
termcolor              1.1.0
thinc                  7.4.1
tifffile               2020.10.1
tqdm                   4.51.0
urllib3                1.25.11
wasabi                 0.8.0
Werkzeug               1.0.1
wheel                  0.35.1
wrapt                  1.12.1
zipp                   3.4.0
